requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
python-multipart==0.0.20
pillow==11.3.0
requests==2.32.5
numpy==2.2.6
matplotlib==3.10.7
langchain-openai==1.0.2
langchain-core==1.0.4
openai==2.7.2
python-dotenv==1.2.1
